리뷰반영

+++ src/main/java/roomescape/service/TimeService.java
@@ -3,6 +3,7 @@ package roomescape.service;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDAO;
 import roomescape.domain.Time;
 import roomescape.dao.TimeDAO;
 
@@ -11,6 +12,7 @@ import roomescape.dao.TimeDAO;
 public class TimeService {
 
     private final TimeDAO timeDAO;
+    private final ReservationDAO reservationDAO;
 
     public Time add(Time time) {
         return timeDAO.save(time);
@@ -21,6 +23,9 @@ public class TimeService {
     }
 
     public void delete(Long id) {
+        if (reservationDAO.findByTimeId(id)) {
+            throw new IllegalStateException("해당 시간대는 예약이 존재하여 삭제할 수 없습니다.");
+        }
         timeDAO.deleteById(id);
     }
 }
